requirements.txt
streamlit
openai-whisper
jiwer
ffmpeg-python
pydub
wikipedia
sentence-transformers
faiss-cpu
numpy
transformers
sentencepiece
torch